- Updating seed script from gulp to npm
