+++ scripts/update-associations.js
